moved aricraft state class
